reorganize file structure, add helpers for tag normalization
